2nd advanced task and readme update

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -43,7 +43,7 @@ int main(void)
 	int i = 0;
 
 	Fibonacci_arr(arr, 1, 50);
-	for (; i < sizeof(arr) / 8; i++)
+	for (; i < (int)sizeof(arr) / 8; i++)
 	{
 		if (i > 9)
 		{
